Tolerate invalid JSON response from EPG API
Quite frequently (several times per day for me) the Timeline view fails to render, because the EPG API returns an invalid JSON body. It appears that there are some strange control characters being returned in the `summary` and `subtitle` fields. This simple patch strips out the ASCII control character range from the API response before attempting to parse the string as JSON. This fixes my issue, and hopefully will be useful to others with the same issue.